backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary>=2.9.9
anthropic>=0.8.0
openai>=1.6.0
sendgrid==6.10.0
python-jose==3.3.0
python-dotenv==1.0.0
pgvector==0.2.4
gunicorn==21.2.0
cryptography>=42.0.0
